+++ xmipp3/protocols/protocol_random_conical_tilt.py
@@ -30,9 +30,9 @@ from os.path import exists
 from pyworkflow.object import String
 from pyworkflow.protocol.constants import LEVEL_ADVANCED
 import pyworkflow.protocol.params as params
-from pyworkflow.em.protocol import ProtInitialVolume
-from pyworkflow.em.data import Volume, SetOfParticles
-from pyworkflow.em.constants import ALIGN_2D
+from pwem.protocols import ProtInitialVolume
+from pwem.objects import Volume, SetOfParticles
+from pwem.constants import ALIGN_2D
 
 import xmippLib
 from xmipp3.base import XmippMdRow
@@ -172,7 +172,7 @@ class XmippProtRCT(ProtInitialVolume):
                 pairRow.setValue(xmippLib.MDL_XCOOR, uCoord.getX())
                 pairRow.setValue(xmippLib.MDL_YCOOR, uCoord.getY())
                 pairRow.setValue(xmippLib.MDL_ENABLED, 1)
-                pairRow.setValue(xmippLib.MDL_ITEM_ID, long(imgId))
+                pairRow.setValue(xmippLib.MDL_ITEM_ID, int(imgId))
                 pairRow.setValue(xmippLib.MDL_REF, 1)
     
                 alignment = img.getTransform()
